interface.h
created the file
added comments
moved over stuff from interface.c
